Fix header name sync — replaced session-based username with DB lookup using userId + role

+++ src/main/java/vv/pms/project/Project.java
@@ -54,7 +54,7 @@ public class Project {
     }
 
     public void setId(long id) {
-        this.id = id;
+        this.id = (Long) id;
     }
 
     public String getTitle() {
